pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "minidispatch"
version = "0.1.0"
description = "Simplified job dispatch library with recovery, CUDA multi-device, and tqdm progress"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "LGPL-3.0"}
dependencies = ["pyyaml", "tqdm>=4.9.0"]

[project.optional-dependencies]
ssh = ["paramiko"]
dev = ["pytest", "ruff", "pyright", "paramiko"]

[project.scripts]
minidispatch = "minidispatch.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pyright]
pythonVersion = "3.9"
